// 5-Design-ParkingLot/code/src/enum.ts
export enum VEHICLE_TYPE {
  BIKE = "BIKE",
  CAR = "CAR",
  BUS = "BUS",
}

export enum SPOT_SIZE {
  SMALL = "SMALL",
  MEDIUM = "MEDIUM",
  LARGE = "LARGE",
}

export enum SPOT_STATUS {
  AVAIALBLE = "AVAIALBLE",
  BLOCKED = "BLOCKED",
  BOOKED = "BOOKED",
}

export enum TICKET_STATUS {
  PENDING = "PENDING",
  PAID = "PAID",
  EXPIRED = "EXPIRED",
}

export enum PAYEMNT_METHOD {
  CREDIT = "CREDIT",
  DEBIT = "DEBIT",
  UPI = "UPI",
}

export enum PAYEMNT_STATUS {
  SUCCESS = "SUCCESS",
  FAILED = "FAILED",
}